+++ src/visualization/printer_impl.h
@@ -12,10 +12,10 @@ struct printer_impl : printer_type
     stream_type& global_stream(point_2f const & pt);
 
     printer_impl(   boost::function<void (int, int, const char *)>          const & draw_string_corner,
-                    boost::function<void (double, double, const char *)>    const & draw_string_global);
+                    boost::function<void (float, float, const char *)>    const & draw_string_global);
 
 private:
-    boost::function<void (double, double, const char *)>    draw_string_global_;
+    boost::function<void (float, float, const char *)>    draw_string_global_;
     int corner_stream_height_indent_;
 
     boost::scoped_ptr<stream_type> corner_stream_;
